src/server/requirements/base.txt
# Django
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
